Enable lots of warnings, and fix the things that they warn about.
Notably, this removes
  o  two implicit uses of self inside blocks in DNSQuery;
  o  a suspicious conversion of a unichar to unsigned char
     inside stringByURLEscaping.

+++ Tidbits/InlineTiming.m
@@ -52,7 +52,7 @@
     NSUInteger max_line = 0;
     NSMutableString* str = [NSMutableString string];
     NSTimeInterval previous = times[0];
-    for (int i = 1; i < count; i++) {
+    for (NSUInteger i = 1; i < count; i++) {
         NSTimeInterval time = times[i];
         NSTimeInterval delta = time - previous;
         [str appendFormat:@" %lf", delta];
